효효효

+++ mbtisland/src/main/java/com/kosta/mbtisland/entity/FileVo.java
@@ -39,7 +39,7 @@ public class FileVo {
 	@Column
 	private String fileType;
 	@Column
-	private Integer fileSize;
+	private Long fileSize; // 타입 Integer->Long 변경
 	@Column
 	@CreationTimestamp
 	private Timestamp uploadDate;
